+++ src/message.rs
@@ -1,5 +1,13 @@
+// use std::path::PathBuf;
+use uuid::Uuid;
+
 #[derive(Debug, Clone)]
 pub enum Message {
-    FolderPathChanged(String),  // フォルダパスの変更
-    GenerateTree,               // ツリー生成ボタンが押された
+    FolderPathChanged(String),
+    OpenFolderDialog,
+    // FolderSelected(Option<PathBuf>),
+    ApplyFilter, // フィルタリングを適用
+    ToggleNodeCheck(Uuid, bool), // ノードのUUIDとチェック状態
+    ShowFilterModal(bool), // フィルタモーダルの表示/非表示を切り替える
+    CopyToClipboard,
 }
